Add reco params to Addon
